Add better workaround for unicode characters in path/file names on Windows
Note: There is still an issue in Python 2's subprocess module that wil fail to properly send
path/file names with unicode characters to avconv/ffmpeg that is out of my control. Any file
with unicode characters in them that require transcoding will fail on the transcoding step
with a message saying as such. This includes non-mp3 files as well as mp3 files when
scan-and-match is enabled.